status-crud

+++ API/Handlers/Products/Delete.cs
@@ -1,6 +1,8 @@
 using System;
+using System.Net;
 using System.Threading;
 using System.Threading.Tasks;
+using API.Infrastructure.Errors;
 using API.Models;
 using MediatR;
 
@@ -25,6 +27,8 @@ namespace API.Handlers.Products
             public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
             {
                 Product product = await context.products.FindAsync(request.product_id);
+
+                if(product == null) throw new RestException(HttpStatusCode.NotFound, new {product = "Not found"});
 
                 context.products.Remove(product);
